requirements.txt
flask
sqlalchemy
psycopg2
python-dotenv
schedule